requirements.txt
alembic~=1.16.4
fastapi~=0.116.1
httpx~=0.28.1
pyyaml~=6.0.2
sqlalchemy~=2.0.43
uvicorn~=0.35.0
pyjwt~=2.10.0
pyzmq~=27.0.2
pydantic-settings~=2.10.1
